// src/app/languages/interface/auth/register.interface.ts
export interface IRegister {
    fullName: string;
    enterFirstName: string;
    lastName: string;
    enterLastName: string;
    mobileNo: string;
    enterMobileNo: string;
    password: string;
    enterPassword: string;
    signUp: string;
    loading: string;
    backTo: string;
    login: string;
}